+++ lisp/_prepare.el
@@ -1681,7 +1681,8 @@ For arguments MSG and ARGS, please see function `eask-print' for the detials."
 
 For arguments MSG and ARGS, please see function `eask--format-paint-kwds' for
 the detials."
-  (message (apply #'eask--format-paint-kwds msg args)))
+  (apply #'eask-write msg args)
+  (eask-princ "\n" t))
 
 (defun eask-write (msg &rest args)
   "Like the function `eask-msg' but without newline at the end.